Fix distribution parsing
The Distribution line in changelogs is actually a space separated list
of distributions where this change should be uploaded to.

Also rename the field to Dists to users of this package can find their
usage of this buggy field in their programs at compile time and thus fix
the cascading bugs caused by that parse error in their programs easily.